Move router endpoints tests inside router module
